gr: fix searching from fm when currrent dir != CWD
In case nnn entered dir other than CWD, then grep will still search
in CWD if call it via ';s'.

The problem was that gr/main sends 'create ...' cmd to dvtm and exits
after, so dvtm can't identify the CWD of client's pid, because it is
gone.

So add '-c' option to 'wm create' which allows to specify particular
CWD to use.